requirements.txt
python-dotenv
flask[async]
flask
requests
typing
redis
cerberus